GCO-2353 Suggestions conditioneel in SearchInput.test gebakken.
