+++ src/components/TupleSearch.tsx
@@ -2,15 +2,12 @@ import React from 'react';
 import querystring from 'querystring';
 import TupleContext from './TupleContext';
 import TupleResult from '../lib/TupleResult';
-import SearchInputWithSuggestions from './SearchInputWithSuggestions';
-import SearchInput from './SearchInput';
+import SearchInput, { SuggestProps } from './SearchInput';
 
 type Props = {
   endpoint: string;
   queryParam?: string;
-  suggestEndpoint?: string,
-  suggestQueryParam?: string,
-  suggestionDelay?:number,
+  suggest?: SuggestProps;
   defaultValue?: string;
   children: (data: TupleResult) => JSX.Element,
 };
@@ -38,20 +35,9 @@ class TupleSearch extends React.Component<Props, State> {
   })
 
   render() {
-    const { suggestEndpoint, suggestQueryParam, suggestionDelay } = this.props;
-
-    let searchInput;
-    if (suggestEndpoint) {
-      searchInput = <SearchInputWithSuggestions onInputChange={this.handleInputChange}
-                                                endpoint={suggestEndpoint}
-                                                queryParam={suggestQueryParam} searchDelay={suggestionDelay}/>;
-    } else {
-      searchInput = <SearchInput onInputChange={this.handleInputChange} />;
-    }
-
     return (
       <div>
-        {searchInput}
+        <SearchInput onInputChange={this.handleInputChange} suggest={this.props.suggest} />
         {this.state.query !== undefined && (
           <div style={{ marginTop: 15 }}>
             <TupleContext src={this.buildUrl()} >
