+++ src/examples/index.html
@@ -74,29 +74,6 @@
       content: "";
     }
 
-    .no-suggestions {
-      color: #999;
-      padding: 0.5rem;
-    }
-
-    .suggestions {
-      border: 1px solid #bebebe;
-      list-style: none;
-      margin-top: 10px;
-      overflow-y: auto;
-      padding-left: 0;
-    }
-
-    .suggestions li {
-      padding: 0.5rem;
-    }
-
-    .suggestion-active,
-    .suggestions li:hover {
-      background-color: #c6c6c6;
-      cursor: pointer;
-    }
-
   </style>
 </head>
 <body>
